Address warnings generated with /Wall

+++ FlexEngine/src/Audio/AudioManager.cpp
@@ -132,7 +132,7 @@ namespace flex
 		i32 format = AL_FORMAT_STEREO8;
 		i32 sampleRate = 44100;
 		i32 size = (i32)(sampleRate * length);
-		u8* data = (u8*)malloc(size);
+		u8* data = (u8*)malloc((u32)size);
 
 		for (i32 i = 0; i < size; ++i)
 		{
